pyproject.toml
[project]
name = "thesis-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.36",
    "coverage>=7.9.1",
    "databricks-sdk>=0.56.0",
    "isort>=6.0.1",
    "mlflow>=3.1.0",
    "optuna>=4.3.0",
    "pandas>=2.3.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.13",
    "scikit-learn>=1.7.0",
    "xgboost>=3.0.2",
]


[tool.isort]
profile = "black"
